requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
dj-database-url==0.5.0
Django==3.1.5
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.1.0
psycopg2==2.9.1
pycparser==2.20
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0